+++ rio/src/RIO/Vector/Boxed.hs
@@ -21,7 +21,7 @@ module RIO.Vector.Boxed
   , (Data.Vector.!?)
 
   -- ** Extracting subvectors
-  , Data.Vector.slice
+  , V.sliceMaybe
   , Data.Vector.take
   , Data.Vector.drop
   , Data.Vector.splitAt
@@ -222,3 +222,4 @@ module RIO.Vector.Boxed
   ) where
 
 import qualified Data.Vector
+import qualified RIO.Vector as V