+++ rio/src/RIO/Vector/Storable.hs
@@ -22,7 +22,7 @@ module RIO.Vector.Storable
   , (Data.Vector.Storable.!?)
 
   -- ** Extracting subvectors
-  , Data.Vector.Storable.slice
+  , V.sliceMaybe
   , Data.Vector.Storable.take
   , Data.Vector.Storable.drop
   , Data.Vector.Storable.splitAt
@@ -188,3 +188,4 @@ module RIO.Vector.Storable
   ) where
 
 import qualified Data.Vector.Storable
+import qualified RIO.Vector as V