+++ rio/src/RIO/Vector/Unboxed.hs
@@ -22,7 +22,7 @@ module RIO.Vector.Unboxed
   , (Data.Vector.Unboxed.!?)
 
   -- ** Extracting subvectors
-  , Data.Vector.Unboxed.slice
+  , V.sliceMaybe
   , Data.Vector.Unboxed.take
   , Data.Vector.Unboxed.drop
   , Data.Vector.Unboxed.splitAt
@@ -211,3 +211,4 @@ module RIO.Vector.Unboxed
   ) where
 
 import qualified Data.Vector.Unboxed
+import qualified RIO.Vector as V