+++ rio/ChangeLog.md
@@ -1,5 +1,10 @@
 # Changelog for rio
 
+## 0.2.0.0
+
+* Move vector `slice` functions to `Partial` modules and add `RIO.Vector.sliceMaybe` function.
+* require `vector-0.12.1`, dropping support for Stackage LTS 13.x ([why](https://github.com/haskell/vector/issues/276#issuecomment-1133765595))
+
 ## 0.1.25.0
 
 * Prevent warnings when compiling with newer GHC versions